Rendering and updating a list
Things to note, we used the spread operator on the persons state because if we use the splice on the reference to the state, we will be actually updating the state directly, which should be immutable, that, is a mistake.

We should always update the state using the methods provided by react, otherwise unpredictable behaviour is expected, and it is considereded a bad practie.

Furthermore, even if persons is declared as a constant, splice can modify our entity, since persons actually holds the reference to the array. And that behaviour is to be expected of the javascript language.